Extract UI into separate repo
In order to properly distinguish between Thesio (essentialy just a UI for
Thesis Hub), I decided to leave only a skeleton of the UI in its former
place and let Thesio UI to be developed outside this repository.
Thesio contains a lot of site specific and unconfigurable UI elemtents.
Once Thesio is implemented there is a chance we can port som of the UI back.